+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ## Added
 * Added `getInvestorsAt` which returns the investors (non-zero balances) at a particular checkpoint
-* Added `iterateInvestors` to allow an arbitrary callback to iterate over a large list of investors without hitting block gas limits.
+* Added `iterateInvestors` to allow an a subset of `investors` to be returned from the security token in case `investors` is large.
 * `ChangeModuleBudget` in SecurityToken now takes in the change in budget rather than absoulte budget. Added boolean paramter to decide increase/decrease.
 * Added `createCheckpoint() withPerm(CHECKPOINT)` to dividend checkpoint modules.
 * Added `deleteDelegate()` to general permission manager. It will disable the delegate but not delete the perms.